+++ workspaces/arazzo/arazzo-designer-extension/src/constants/index.ts
@@ -17,12 +17,12 @@
  */
 
 export const COMMANDS = {
-    PROJECT_EXPLORER: "APIDesigner.project-explorer",
-    PROJECT_EXPLORER_REFRESH: "APIDesigner.project-explorer.refresh",
-    OPEN_WELCOME: "APIDesigner.openAPIDesigner",
-    CREATE_OPENAPI_FILE: "APIDesigner.createOpenAPIFile",
-    CREATE_ARAZZO_FILE: "APIDesigner.createArazzoFile",
-    SHOW_CODE: "APIDesigner.showCode",
+    PROJECT_EXPLORER: "ArazzoDesigner.project-explorer",
+    PROJECT_EXPLORER_REFRESH: "ArazzoDesigner.project-explorer.refresh",
+    OPEN_WELCOME: "ArazzoDesigner.openAPIDesigner",
+    CREATE_OPENAPI_FILE: "ArazzoDesigner.createOpenAPIFile",
+    CREATE_ARAZZO_FILE: "ArazzoDesigner.createArazzoFile",
+    SHOW_CODE: "ArazzoDesigner.showCode",
     // LSP Code Lens commands
     ARAZZO_VISUALIZE: "arazzo.visualize",
     ARAZZO_OPEN_DESIGNER: "arazzo.openDesigner"